+++ src/component/ui/SelectProps.tsx
@@ -7,24 +7,28 @@ import {
   SelectValue,
 } from '@/components/ui/select';
 
-function SelectProps({
-  value,
-  onChange,
-}: {
+interface SelectProps {
   value: string;
   onChange: (newValue: string) => void;
-}) {
+  disabled?: boolean;
+  className: string;
+}
+
+function SelectProps({ value, onChange, disabled, className }: SelectProps) {
   return (
     <>
-      <Select onValueChange={onChange} value={value}>
-        <SelectTrigger id="category" className="w-full h-14">
+      <Select onValueChange={onChange} value={value} disabled={disabled}>
+        <SelectTrigger
+          id="category"
+          className={`w-full h-14 ${className || ''}`}
+        >
           <SelectValue placeholder="카테고리를 선택해 주세요" />
         </SelectTrigger>
         <SelectContent>
           <SelectGroup>
-            <SelectItem value="LARGE">Large</SelectItem>
-            <SelectItem value="MEDIUM">Medium</SelectItem>
-            <SelectItem value="SMALL">Small</SelectItem>
+            <SelectItem value="LARGE">LARGE</SelectItem>
+            <SelectItem value="MEDIUM">MEDIUM</SelectItem>
+            <SelectItem value="SMALL">SMALL</SelectItem>
           </SelectGroup>
         </SelectContent>
       </Select>
